bugsage-api/requirements.txt
fastapi
uvicorn
langchain
openai
python-dotenv
pinecone
sentence-transformers